Completely rewrote day04
Using the TDD strategy that Rick showed me from Day 5 , I went back and started on day 4 from scratch and built out the rspecs and the code back and forth. For the Part 2 I knew I wanted something that wasn't trying to use recursion into the same array. Rewrote it so it was a simple hash map and now it is MUCH faster. Both Part 1 and Part 2 answers were accepted.